// Frontend-IHM-Budget/src/model/reference.ts
export type ReferenceType = {
  year: number;
  costCenter: string;
  department: string;
  activity: string;
  subActivity: string;
  catPlan: string;
  client: string;
  budgetType: string;
  businessLine: string;
  syntheticMapping: string;
  billingKey: string;
  businessResponsible: string;
  comexMember: string;
  active: boolean;
};